[ms-5] orderRepository에 동적 쿼리 추가 및 테스트 추가

+++ src/main/kotlin/com/narea/mall/dto/ItemDto.kt
@@ -17,16 +17,16 @@ data class ItemCreateRequest (
     @field:NotBlank(message = NOT_BE_EMPTY)
     var name: String = StringUtils.EMPTY,
     var description:String = StringUtils.EMPTY,
-    var price:Long = 0,
-    var inventory:Long = 0,
+    var price: Int = 0,
+    var inventory: Int = 0,
     var categoryIds: List<Long> = emptyList()
 )
 
 data class ItemUpdateRequest (
     var name: String? = null,
     var description:String? = null,
-    var price:Long? = null,
-    var inventory:Long? = null,
+    var price: Int? = null,
+    var inventory: Int? = null,
     var categories:List<Long>? = null,
 )
 
@@ -35,8 +35,8 @@ data class ItemResponse (
     var id:Long = 0,
     var name:String = StringUtils.EMPTY,
     var description:String = StringUtils.EMPTY,
-    var price:Long = 0,
-    var inventory:Long = 0,
+    var price: Int = 0,
+    var inventory: Int = 0,
     var files:List<ItemFileResponse> = emptyList(),
     var categoryList:List<Category> = emptyList()
 )
